requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
mysqlclient==2.1.1
asgiref==3.6.0
celery==5.3.0
Pillow==10.0.0
django-cors-headers==3.14.0